Setup Eureka as First Bootstrap
Setup Eureka as First Bootstrap (eureka is the first thing that starts up and then cloud config. The price for using this option is an extra network round trip on start up to locate the service registration. The benefit is that the Config Server can change its co-ordinates, as long as Eureka is a fixed point.